refactor extraction
