+++ package.json
@@ -56,7 +56,6 @@
     "eslint-plugin-compat": "^4.0.2",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsdoc": "^37.9.4",
-    "extract-gfm": "^0.1.0",
     "glob": "^7.2.0",
     "highlight.js": "^11.5.1",
     "http-server": "^14.1.0",
